requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
pyyaml
streamlit
fastapi
uvicorn[standard]
transformers
scikit-learn
